refactor: update auth
