+++ src/auth/index.ts
@@ -1,4 +1,4 @@
-export * from './authentication-provider'
+export * from './auth-provider'
 export * from './session'
 export * from './access-token'
 export * from './account-information'
