+++ src/commands/pdf/post-pdf-template-builder.ts
@@ -2,7 +2,7 @@ import { Post } from '@/models/post'
 import { PostFileMapManager } from '@/services/post-file-map'
 import fs from 'fs'
 import { blogSettingsService } from '@/services/blog-settings.service'
-import { accountManager } from '@/authentication/account-manager'
+import { accountManager } from '@/auth/account-manager'
 import { postCategoryService } from '@/services/post-category.service'
 import { PostCategory } from '@/models/post-category'
 import { markdownItFactory } from '@cnblogs/markdown-it-presets'
